+++ tests/testthat/test-dig_paired_contrasts.R
@@ -75,3 +75,22 @@ test_that("dig_paired_contrasts var", {
     expect_equal(res$yvar, rep("uptake", 9))
     expect_equal(res$p_value, rep(0, 9))
 })
+
+test_that("dig_paired contrasts errors", {
+    d <- data.frame(n = 1:5 / 5, l = TRUE, i = 1:5, s = letters[1:5])
+
+    expect_error(dig_paired_contrasts(x = 1:5),
+                 "`x` must be a matrix or a data frame")
+    expect_error(dig_paired_contrasts(d, condition = n:l),
+                 "All columns selected by `condition` must be logical.")
+    expect_error(dig_paired_contrasts(d, xvars = n, yvars = l),
+                 "All columns selected by `yvars` must be numeric.")
+    expect_error(dig_paired_contrasts(d, xvars = s, yvars = n),
+                 "All columns selected by `xvars` must be numeric.")
+    expect_error(dig_paired_contrasts(d, method = "foo"),
+                 '`method` must be equal to one of: "t", "wilcox", "var".')
+    expect_error(dig_paired_contrasts(d, alternative = "foo"),
+                 '`alternative` must be equal to one of: "two.sided", "less", "greater".')
+    expect_error(dig_paired_contrasts(d, min_length = "x"),
+                 "`min_length` must be an integerish scalar.")
+})